Extract node editor view diff helpers
- Create node-editor-view-diff.js with pure diff/comparison functions
- Export: cloneEditorModelSnapshot, getAddedNodeIds, getRemovedNodeIds,
  getCommonNodeIds, getAddedEdgeIds, getRemovedEdgeIds, shouldRecreateNode,
  sameParams, samePosition
- Import and use helpers in node-editor-view.js
- Add comprehensive test suite verifying helper behavior
- No runtime behavior changes

Enables testing helpers directly without testing NodeEditorView,
improving test maintainability and specificity.
