+++ editor-studio/src/node-editor-view.js
@@ -8,6 +8,17 @@ import {
   translateToMoveNodeOp,
   controlValueToUpdateParamOp
 } from './rete-operation-helpers.js';
+import {
+  cloneEditorModelSnapshot,
+  getAddedNodeIds,
+  getRemovedNodeIds,
+  getCommonNodeIds,
+  getAddedEdgeIds,
+  getRemovedEdgeIds,
+  shouldRecreateNode,
+  sameParams,
+  samePosition
+} from './node-editor-view-diff.js';
 
 const socket = new ClassicPreset.Socket('value');
 
@@ -47,56 +58,7 @@ function createReteNode(editorNode, onControl) {
   return node;
 }
 
-// --- Snapshot / diff helpers ---
-
-function cloneEditorModelSnapshot(editorModel) {
-  return JSON.parse(JSON.stringify(editorModel));
-}
-
-function canPatchEditorModel(previous, next) {
-  if (!previous || !next) return false;
-  return true;
-}
-
-function getAddedNodeIds(previous, next) {
-  return next.order.filter((id) => !previous.nodesById[id]);
-}
-
-function getRemovedNodeIds(previous, next) {
-  return previous.order.filter((id) => !next.nodesById[id]);
-}
-
-function getCommonNodeIds(previous, next) {
-  return next.order.filter((id) => previous.nodesById[id] && next.nodesById[id]);
-}
-
-function getAddedEdgeIds(previous, next) {
-  return Object.keys(next.edgesById || {}).filter((id) => !previous.edgesById?.[id]);
-}
-
-function getRemovedEdgeIds(previous, next) {
-  return Object.keys(previous.edgesById || {}).filter((id) => !next.edgesById?.[id]);
-}
-
-function shouldRecreateNode(previousNode, nextNode) {
-  if (previousNode.type !== nextNode.type) return true;
-  if (previousNode.category !== nextNode.category) return true;
-
-  const prevParamKeys = Object.keys(previousNode.params || {}).sort();
-  const nextParamKeys = Object.keys(nextNode.params || {}).sort();
-
-  if (prevParamKeys.join('\0') !== nextParamKeys.join('\0')) return true;
-
-  return false;
-}
-
-function sameParams(a, b) {
-  return JSON.stringify(a || {}) === JSON.stringify(b || {});
-}
-
-function samePosition(a, b) {
-  return a?.x === b?.x && a?.y === b?.y;
-}
+// --- Connection helpers ---
 
 function findReteConnectionIdByEdgeId(connectionMap, edgeId) {
   for (const [connectionId, mappedEdgeId] of connectionMap.entries()) {
